pyproject.toml
[project]
name = "ngmidbms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "psycopg[binary]>=3.2.3",
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "PyPDF2>=3.0.0",
    "python-docx>=1.1.0",
    "bcrypt>=4.0.0",
    "python-dotenv>=1.0.0",
    "langchain-community>=0.4.1",
    "pypdf>=6.4.0",
]
